\name{newXBSeqDataSetFromHTSeqCount}
\alias{newXBSeqDataSetFromHTSeqCount}
\title{
   Creat a XBSeqDataSet object from output of HESeq
}
\description{The same method is adopted from DESeq. Use this function to start a DESeq analysis if you used htseq-count
to count your reads.}
\usage{
newXBSeqDataSetFromHTSeqCount(sampleTable, directory = ".")
}
\arguments{
  \item{sampleTable}{
     A data frame with three or more columns. Each row describes one sample.
The first column is the sample name, the seond column the file name of
the count file generated by htseq-count, and the remaining columns are
sample meta data. If the meta data consists of only a single column
(i.e., three columns in total), this is used as 'condition' factor. 
}
  \item{directory}{
The directory relative to which the filenames are specified.
}
}
\value{
A XBSeqDataSet object.}
\references{
   See \url{http://www-huber.embl.de/users/anders/HTSeq/} for htseq-count \cr
   Simon Anders, Wolfgang Huber: Differential expression
   analysis for sequence count data. Genome Biology 11 (2010)
   R106, \url{http://dx.doi.org/10.1186/gb-2010-11-10-r106}   
}
\author{
   Yuanhang Liu
}
\seealso{
\code{\link{newXBSeqDataSet}}
}